QNX/BB10: Network, IPC, and Mojo adaptations
Port network stack, IPC channel, and Mojo bindings to QNX. Includes
HTTP header deserialization fix for QNX's non-standard struct layout,
network service threading adaptations, SimpleWatcher tracing for
cross-thread Mojo message delivery, and socket/DNS compatibility shims.

+++ net/dns/host_resolver_system_task.cc
@@ -273,7 +273,7 @@ void HostResolverSystemTask::StartLookupAttempt() {
         base::BindOnce(&HostResolverSystemTask::StartLookupAttempt,
                        weak_ptr_factory_.GetWeakPtr()),
         params_.unresponsive_delay *
-            std::pow(params_.retry_factor, attempt_number_ - 1));
+            std::pow(static_cast<double>(params_.retry_factor), static_cast<double>(attempt_number_ - 1)));
   }
 
   auto lookup_complete_cb =
